misc: formatting changes
